Reorganise board-related files
