+++ src/components/leaderboards/View.vue
@@ -1,12 +1,12 @@
 <script setup lang="ts">
 import { useAsyncState } from '@vueuse/core'
 import { computed, ref } from 'vue'
-import { useApi } from '../composables/useApi'
-import { useAuth } from '../composables/useAuth'
-import { useSessionToken } from '../composables/useSessionToken'
-import { Leaderboards } from '../lib/api/Leaderboards'
-import { ProblemDetails } from '../lib/api/data-contracts'
-import { HttpResponse } from '../lib/api/http-client'
+import { useApi } from '../../composables/useApi'
+import { useAuth } from '../../composables/useAuth'
+import { useSessionToken } from '../../composables/useSessionToken'
+import { Leaderboards } from '../../lib/api/Leaderboards'
+import { ProblemDetails } from '../../lib/api/data-contracts'
+import { HttpResponse } from '../../lib/api/http-client'
 
 const props = defineProps<{
 	id: number
